requirements.txt
requests
pandas
numpy
yfinance
python-dotenv
openpyxl
